Extracted add participant validation logic out to a base component
