+++ AdminWebsite/AdminWebsite/ClientApp/src/app/booking/add-participant-base/add-participant-base.component.spec.ts
@@ -1,0 +1,25 @@
+// import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+// import { AddParticipantBaseComponent } from './add-participant-base.component';
+
+// describe('AddParticipantBaseComponent', () => {
+//   let component: AddParticipantBaseComponent;
+//   let fixture: ComponentFixture<AddParticipantBaseComponent>;
+
+//   beforeEach(async () => {
+//     await TestBed.configureTestingModule({
+//       declarations: [ AddParticipantBaseComponent ]
+//     })
+//     .compileComponents();
+//   });
+
+//   beforeEach(() => {
+//     fixture = TestBed.createComponent(AddParticipantBaseComponent);
+//     component = fixture.componentInstance;
+//     fixture.detectChanges();
+//   });
+
+//   it('should create', () => {
+//     expect(component).toBeTruthy();
+//   });
+// });
