+++ AdminWebsite/AdminWebsite/ClientApp/src/app/booking/add-participant/add-participant.component.ts
@@ -1,21 +1,17 @@
 import { AfterContentInit, AfterViewInit, Component, OnDestroy, OnInit, ViewChild } from '@angular/core';
-import { FormControl, FormGroup, Validators } from '@angular/forms';
 import { Router } from '@angular/router';
 import { Observable, Subscription } from 'rxjs';
 import { PageUrls } from 'src/app/shared/page-url.constants';
 import { Constants } from '../../common/constants';
 import { SanitizeInputText } from '../../common/formatters/sanitize-input-text';
 import { IDropDownModel } from '../../common/model/drop-down.model';
-import { HearingModel } from '../../common/model/hearing.model';
 import { ParticipantModel } from '../../common/model/participant.model';
-import { PartyModel } from '../../common/model/party.model';
 import { BookingService } from '../../services/booking.service';
 import { CaseAndHearingRolesResponse, LinkedParticipantRequest } from '../../services/clients/api-client';
 import { Logger } from '../../services/logger';
 import { SearchService } from '../../services/search.service';
 import { VideoHearingsService } from '../../services/video-hearings.service';
-import { BookingBaseComponentDirective as BookingBaseComponent } from '../booking-base/booking-base.component';
-import { SearchEmailComponent } from '../search-email/search-email.component';
+import { AddParticipantBaseDirective } from 'src/app/booking/add-participant-base/add-participant-base.component'
 import { ParticipantService } from '../services/participant.service';
 import { ParticipantListComponent } from '../participant';
 import { HearingRoles } from '../../common/model/hearing-roles.model';
@@ -26,64 +22,29 @@ import { LinkedParticipantModel, LinkedParticipantType } from 'src/app/common/mo
     templateUrl: './add-participant.component.html',
     styleUrls: ['./add-participant.component.scss']
 })
-export class AddParticipantComponent extends BookingBaseComponent implements OnInit, AfterViewInit, AfterContentInit, OnDestroy {
+export class AddParticipantComponent extends AddParticipantBaseDirective implements OnInit, AfterViewInit, AfterContentInit, OnDestroy {
     constants = Constants;
 
-    emailDisabled = false;
-    participantDetails: ParticipantModel;
     notFound: boolean;
-    hearing: HearingModel;
     titleList: IDropDownModel[] = [];
     roleList: string[];
-    hearingRoleList: string[];
-    caseAndHearingRoles: PartyModel[] = [];
     selectedParticipantEmail: string = null;
-    private role: FormControl;
-    private party: FormControl;
-    private title: FormControl;
-    private firstName: FormControl;
-    private lastName: FormControl;
-    private phone: FormControl;
-    private displayName: FormControl;
-    private companyName: FormControl;
-    private companyNameIndividual: FormControl;
-    private representing: FormControl;
-    isRoleSelected = true;
-    isPartySelected = true;
     isTitleSelected = true;
-    isShowErrorSummary = false;
-    showDetails = false;
     showCancelPopup = false;
     showConfirmationPopup = false;
     attemptingDiscardChanges = false;
     confirmationMessage: string;
     showConfirmationRemoveParticipant = false;
     removerFullName: string;
-    displayNextButton = true;
-    displayAddButton = false;
-    displayClearButton = false;
-    displayUpdateButton = false;
-    displayErrorNoParticipants = false;
     localEditMode = false;
     isExistingHearing: boolean;
     isAnyParticipants: boolean;
     existingPerson: boolean;
-    existingParticipant: boolean;
-    isRepresentative = false;
     bookingHasParticipants: boolean;
-    existingPersonEmails: string[] = [];
     $subscriptions: Subscription[] = [];
 
-    private interpreterFor: FormControl;
     interpreteeList: ParticipantModel[] = [];
-    isInterpreter = false;
     showConfirmRemoveInterpretee = false;
-    interpreterSelected = false;
-    errorAlternativeEmail = false;
-    errorJohAccountNotFound = false;
-    errorJudiciaryAccount = false;
-    errorNotFoundJohEmail = false;
-    @ViewChild(SearchEmailComponent) searchEmail: SearchEmailComponent;
 
     @ViewChild(ParticipantListComponent, { static: true })
     participantsListComponent: ParticipantListComponent;
@@ -197,53 +158,7 @@ export class AddParticipantComponent extends BookingBaseComponent implements OnI
     }
 
     initializeForm() {
-        this.role = new FormControl(this.constants.PleaseSelect, [
-            Validators.required,
-            Validators.pattern(this.constants.PleaseSelectPattern)
-        ]);
-        this.party = new FormControl(this.constants.PleaseSelect, [
-            Validators.required,
-            Validators.pattern(this.constants.PleaseSelectPattern)
-        ]);
-        this.title = new FormControl(this.constants.PleaseSelect);
-        this.firstName = new FormControl('', [
-            Validators.required,
-            Validators.pattern(Constants.TextInputPatternName),
-            Validators.maxLength(255)
-        ]);
-        this.lastName = new FormControl('', [
-            Validators.required,
-            Validators.pattern(Constants.TextInputPatternName),
-            Validators.maxLength(255)
-        ]);
-        this.phone = new FormControl('', [Validators.required, Validators.pattern(Constants.PhonePattern)]);
-        this.displayName = new FormControl('', [
-            Validators.required,
-            Validators.pattern(Constants.TextInputPatternName),
-            Validators.maxLength(255)
-        ]);
-        this.companyName = new FormControl('');
-        this.companyNameIndividual = new FormControl('', [Validators.pattern(Constants.TextInputPattern), Validators.maxLength(255)]);
-        this.representing = new FormControl('', [Validators.pattern(Constants.TextInputPattern), Validators.maxLength(255)]);
-        this.interpreterFor = new FormControl(this.constants.PleaseSelect, [
-            Validators.required,
-            Validators.pattern(this.constants.PleaseSelectPattern)
-        ]);
-
-        this.form = new FormGroup({
-            role: this.role,
-            party: this.party,
-            title: this.title,
-            firstName: this.firstName,
-            lastName: this.lastName,
-            phone: this.phone,
-            displayName: this.displayName,
-            companyName: this.companyName,
-            companyNameIndividual: this.companyNameIndividual,
-            representing: this.representing,
-            interpreterFor: this.interpreterFor
-        });
-
+        this.initialiseForm();
         const self = this;
         this.$subscriptions.push(
             this.form.valueChanges.subscribe(result => {
@@ -314,89 +229,13 @@ export class AddParticipantComponent extends BookingBaseComponent implements OnI
 
     setupRoles(data: CaseAndHearingRolesResponse[]) {
         this.caseAndHearingRoles = this.participantService.mapParticipantsRoles(data);
-        this.roleList = this.caseAndHearingRoles.filter(x => x.name !== 'Judge').map(x => x.name);
+        this.roleList = this.caseAndHearingRoles.filter(x => x.name !== 'Judge' && x.name !== 'Staff Member').map(x => x.name);
         this.roleList.unshift(this.constants.PleaseSelect);
         this.caseAndHearingRoles.forEach(x => {
             this.setupHearingRoles(x.name);
         });
     }
 
-    setupHearingRoles(caseRoleName: string) {
-        const list = this.caseAndHearingRoles.find(x => x.name === caseRoleName && x.name !== 'Judge');
-        this.hearingRoleList = list ? list.hearingRoles.map(x => x.name) : [];
-        this.updateHearingRoleList(this.hearingRoleList);
-        if (!this.hearingRoleList.find(s => s === this.constants.PleaseSelect)) {
-            this.hearingRoleList.unshift(this.constants.PleaseSelect);
-        }
-    }
-
-    public getParticipant(participantDetails: ParticipantModel) {
-        if (!this.validateJudgeAndJohMembers()) {
-            this.searchEmail.isErrorEmailAssignedToJudge = true;
-            this.errorAlternativeEmail = true;
-            return;
-        }
-        this.errorAlternativeEmail = false;
-        this.errorJohAccountNotFound = false;
-        this.displayErrorNoParticipants = false;
-        this.displayAdd();
-        this.enableFields();
-        this.participantDetails = Object.assign({}, participantDetails);
-
-        if (participantDetails.is_exist_person) {
-            this.disableLastFirstNames();
-            this.emailDisabled = true;
-            this.existingPersonEmails.push(participantDetails.email);
-        }
-        this.existingParticipant = participantDetails.id && participantDetails.id.length > 0;
-        if (this.existingParticipant) {
-            this.disableCaseAndHearingRoles();
-            this.disableLastFirstNames();
-        }
-        // if it's added in the existing hearing participant, then allowed all fields to edit.
-        this.resetPartyAndRole();
-
-        this.isRepresentative = this.isRoleRepresentative(this.participantDetails.hearing_role_name, this.party.value);
-
-        this.form.setValue({
-            party: this.participantDetails.case_role_name,
-            role: this.participantDetails.hearing_role_name,
-            title: this.participantDetails.title === undefined ? this.constants.PleaseSelect : this.participantDetails.title,
-            firstName: this.participantDetails.first_name,
-            lastName: this.participantDetails.last_name,
-            phone: this.participantDetails.phone || '',
-            displayName: this.participantDetails.display_name || '',
-            companyName: this.participantDetails.company || '',
-            companyNameIndividual: this.participantDetails.company || '',
-            representing: this.participantDetails.representee || '',
-            interpreterFor: this.setInterpretee(this.participantDetails) || this.constants.PleaseSelect
-        });
-
-        setTimeout(() => {
-            this.form.get('role').setValue(this.participantDetails.hearing_role_name);
-            this.roleSelected();
-        }, 500);
-    }
-
-    resetPartyAndRole() {
-        if (this.participantDetails.case_role_name) {
-            this.setupHearingRoles(this.participantDetails.case_role_name);
-        }
-        if (
-            this.isPartySelected &&
-            !this.existingParticipant &&
-            (!this.participantDetails.case_role_name || this.participantDetails.case_role_name.length === 0)
-        ) {
-            this.participantDetails.case_role_name = this.party.value;
-        }
-        if (
-            this.isRoleSelected &&
-            !this.existingParticipant &&
-            (!this.participantDetails.hearing_role_name || this.participantDetails.hearing_role_name.length === 0)
-        ) {
-            this.participantDetails.hearing_role_name = this.role.value;
-        }
-    }
 
     notFoundParticipant() {
         this.logger.warn(`${this.loggerPrefix} Participant not found.`);
@@ -410,33 +249,6 @@ export class AddParticipantComponent extends BookingBaseComponent implements OnI
         }
     }
 
-    emailChanged() {
-        if (!this.validateJudgeAndJohMembers()) {
-            this.searchEmail.isErrorEmailAssignedToJudge = true;
-            this.errorAlternativeEmail = true;
-            this.errorJohAccountNotFound = false;
-            return;
-        }
-        this.errorAlternativeEmail = false;
-        this.errorJohAccountNotFound = false;
-
-        if (this.form.valid && this.validEmail()) {
-            this.disableCaseAndHearingRoles();
-            if (this.editMode) {
-                this.displayNext();
-            } else {
-                this.displayAdd();
-            }
-        }
-    }
-
-    private displayAdd() {
-        this.displayNextButton = false;
-        this.displayClearButton = true;
-        this.displayAddButton = true;
-        this.displayUpdateButton = false;
-    }
-
     private displayUpdate() {
         this.displayNextButton = false;
         this.displayClearButton = true;
@@ -444,60 +256,11 @@ export class AddParticipantComponent extends BookingBaseComponent implements OnI
         this.displayAddButton = false;
     }
 
-    private displayNext() {
-        this.displayNextButton = true;
-        this.displayClearButton = false;
-        this.displayAddButton = false;
-        this.displayUpdateButton = false;
-    }
-
     private displayClear() {
         this.displayNextButton = false;
         this.displayClearButton = true;
         this.displayAddButton = false;
         this.displayUpdateButton = false;
-    }
-
-    get firstNameInvalid() {
-        return this.firstName.invalid && (this.firstName.dirty || this.firstName.touched || this.isShowErrorSummary);
-    }
-
-    get lastNameInvalid() {
-        return this.lastName.invalid && (this.lastName.dirty || this.lastName.touched || this.isShowErrorSummary);
-    }
-
-    get phoneInvalid() {
-        return this.phone.invalid && (this.phone.dirty || this.phone.touched || this.isShowErrorSummary);
-    }
-
-    get partyInvalid() {
-        return this.party.invalid && (this.party.dirty || this.party.touched || this.isShowErrorSummary);
-    }
-
-    get roleInvalid() {
-        return this.role.invalid && (this.role.dirty || this.role.touched || this.isShowErrorSummary);
-    }
-
-    get displayNameInvalid() {
-        return (
-            (this.displayName.invalid && (this.displayName.dirty || this.displayName.touched || this.isShowErrorSummary)) ||
-            (this.displayName.touched && this.displayName.value === '')
-        );
-    }
-
-    get representeeInvalid() {
-        return this.representing.invalid && (this.representing.dirty || this.representing.touched || this.isShowErrorSummary);
-    }
-
-    get companyInvalid() {
-        return this.companyName.invalid && (this.companyName.dirty || this.companyName.touched || this.isShowErrorSummary);
-    }
-
-    get companyIndividualInvalid() {
-        return (
-            this.companyNameIndividual.invalid &&
-            (this.companyNameIndividual.dirty || this.companyNameIndividual.touched || this.isShowErrorSummary)
-        );
     }
 
     partySelected() {
@@ -511,53 +274,9 @@ export class AddParticipantComponent extends BookingBaseComponent implements OnI
         this.validateJudiciaryEmailAndRole();
     }
 
-    roleSelected() {
-        this.isRoleSelected = this.role.value !== this.constants.PleaseSelect;
-        if (!this.isRoleRepresentative(this.role.value, this.party.value)) {
-            this.companyName.clearValidators();
-            this.representing.clearValidators();
-
-            this.companyName.updateValueAndValidity();
-            this.representing.updateValueAndValidity();
-
-            this.companyName.setValue('');
-            this.representing.setValue('');
-        } else {
-            this.companyName.setValidators([
-                Validators.required,
-                Validators.pattern(Constants.TextInputPattern),
-                Validators.maxLength(255)
-            ]);
-            this.representing.setValidators([
-                Validators.required,
-                Validators.pattern(Constants.TextInputPattern),
-                Validators.maxLength(255)
-            ]);
-
-            this.companyName.updateValueAndValidity();
-            this.representing.updateValueAndValidity();
-
-            this.companyNameIndividual.setValue('');
-        }
-        this.showDetails = true;
-        this.isRepresentative = this.isRoleRepresentative(this.role.value, this.party.value);
-        this.setInterpreterForValidation();
-    }
 
     titleSelected() {
         this.isTitleSelected = this.title.value !== this.constants.PleaseSelect;
-    }
-
-    validEmail() {
-        return this.showDetails && this.searchEmail ? this.searchEmail.validateEmail() : true;
-    }
-
-    validateJudgeAndJohMembers(): boolean {
-        if (this.hearing?.participants.length) {
-            const judge = this.hearing.participants.find(x => x.is_judge);
-            return this.searchEmail?.email !== judge?.username;
-        }
-        return true;
     }
 
     validateJudiciaryEmailAndRole() {
@@ -585,6 +304,7 @@ export class AddParticipantComponent extends BookingBaseComponent implements OnI
 
     saveParticipant() {
         this.actionsBeforeSave();
+        
         if (
             this.form.valid &&
             this.validEmail() &&
@@ -607,6 +327,7 @@ export class AddParticipantComponent extends BookingBaseComponent implements OnI
             }
 
             this.mapParticipant(newParticipant);
+            
             if (!this.participantService.checkDuplication(newParticipant.email, this.hearing.participants)) {
                 this.addLinkedParticipant(newParticipant);
 
@@ -857,6 +578,7 @@ export class AddParticipantComponent extends BookingBaseComponent implements OnI
             firstName: '',
             lastName: '',
             phone: '',
+            email: '',
             displayName: '',
             companyName: '',
             companyNameIndividual: '',
@@ -930,24 +652,6 @@ export class AddParticipantComponent extends BookingBaseComponent implements OnI
         window.document.getElementById(fragment).parentElement.parentElement.scrollIntoView();
     }
 
-    disableLastFirstNames() {
-        this.form.get('lastName').disable();
-        this.form.get('firstName').disable();
-    }
-
-    disableCaseAndHearingRoles() {
-        this.form.get('party').disable();
-        this.form.get('role').disable();
-    }
-
-    enableFields() {
-        this.emailDisabled = false;
-        this.form.get('lastName').enable();
-        this.form.get('firstName').enable();
-        this.form.get('party').enable();
-        this.form.get('role').enable();
-    }
-
     firstNameOnBlur() {
         const text = SanitizeInputText(this.firstName.value);
         this.firstName.setValue(text);
@@ -987,19 +691,6 @@ export class AddParticipantComponent extends BookingBaseComponent implements OnI
         });
     }
 
-    isRoleRepresentative(hearingRole: string, party: string): boolean {
-        const partyHearingRoles = this.caseAndHearingRoles.find(
-            x => x.name === party && x.name !== 'Judge' && x.hearingRoles.find(y => y.name === hearingRole)
-        );
-
-        if (!partyHearingRoles) {
-            return false;
-        }
-
-        const findHearingRole = partyHearingRoles.hearingRoles.find(x => x.name === hearingRole);
-        return findHearingRole && findHearingRole.userRole === 'Representative';
-    }
-
     handleContinueRemoveInterpreter() {
         this.showConfirmRemoveInterpretee = false;
         this.removeInterpreteeAndInterpreter();
@@ -1034,46 +725,7 @@ export class AddParticipantComponent extends BookingBaseComponent implements OnI
         this.interpreteeList.unshift(interpreteeModel);
     }
 
-    private setInterpreterForValidation() {
-        if (this.isRoleInterpreter(this.role.value)) {
-            this.interpreterFor.setValidators([Validators.required, Validators.pattern(Constants.PleaseSelectPattern)]);
-            this.interpreterFor.updateValueAndValidity();
-            this.isInterpreter = true;
-        } else {
-            this.interpreterFor.clearValidators();
-            this.interpreterFor.updateValueAndValidity();
-            this.interpreterFor.setValue(Constants.PleaseSelect);
-            this.isInterpreter = false;
-        }
-    }
 
-    private isRoleInterpreter(hearingRole: string): boolean {
-        return hearingRole.toLowerCase() === HearingRoles.INTERPRETER.toLowerCase();
-    }
-
-    private hearingHasAnInterpreter(): boolean {
-        const hearingHasInterpreter = this.hearing.participants.some(
-            p => p.hearing_role_name?.toLowerCase() === HearingRoles.INTERPRETER.toLowerCase()
-        );
-        return hearingHasInterpreter;
-    }
-
-    private hearingHasInterpretees(): boolean {
-        const notAllowedInterpreter = [HearingRoles.INTERPRETER.toLowerCase(), HearingRoles.OBSERVER.toLowerCase()];
-        const hearingHasInterpretees = this.hearing.participants.some(
-            p => p.user_role_name === 'Individual' && !notAllowedInterpreter.includes(p.hearing_role_name.toLowerCase())
-        );
-        return hearingHasInterpretees;
-    }
-
-    private updateHearingRoleList(hearingRoleList: string[]) {
-        // hide the interpreter value if participant list is empty or participant list has an interpreter.
-        if (this.hearingHasAnInterpreter() || !this.hearingHasInterpretees()) {
-            if (!this.interpreterSelected) {
-                this.hearingRoleList = this.hearingRoleList.filter(item => item.toLowerCase() !== HearingRoles.INTERPRETER);
-            }
-        }
-    }
 
     private removeInterpreteeAndInterpreter() {
         // check if participant details were populated, if yes then clean form.
@@ -1126,16 +778,5 @@ export class AddParticipantComponent extends BookingBaseComponent implements OnI
     private getInterpretee(email: string): string {
         const interpretee = this.hearing.participants.find(p => p.email === email);
         return interpretee ? interpretee.email : '';
-    }
-
-    private setInterpretee(participant: ParticipantModel): string {
-        let interpreteeEmail = '';
-        if (participant.interpreterFor) {
-            interpreteeEmail = participant.interpreterFor;
-        } else if (participant.linked_participants && participant.linked_participants.length > 0) {
-            const interpretee = this.hearing.participants.find(p => p.id === participant.linked_participants[0].linkedParticipantId);
-            interpreteeEmail = interpretee ? interpretee.email : '';
-        }
-        return interpreteeEmail;
     }
 }
