+++ AdminWebsite/AdminWebsite/ClientApp/src/app/booking/add-staff-member/add-staff-member.component.spec.ts
@@ -1,4 +1,12 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
+import { Router } from '@angular/router';
+import { RouterTestingModule } from '@angular/router/testing';
+import { BookingService } from 'src/app/services/booking.service';
+import { Logger } from 'src/app/services/logger';
+import { VideoHearingsService } from 'src/app/services/video-hearings.service';
+import { provide } from 'src/app/testing/helpers/jasmine.helpers';
+import { AddParticipantBaseDirective } from '../add-participant-base/add-participant-base.component';
 
 import { AddStaffMemberComponent } from './add-staff-member.component';
 
@@ -6,12 +14,31 @@ describe('AddStaffMemberComponent', () => {
   let component: AddStaffMemberComponent;
   let fixture: ComponentFixture<AddStaffMemberComponent>;
 
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      declarations: [ AddStaffMemberComponent ]
+  let bookingServiceSpy: jasmine.SpyObj<BookingService>;
+  let videoHearingsServiceSpy: jasmine.SpyObj<VideoHearingsService>;
+  let routerSpy: jasmine.SpyObj<Router>;
+  let loggerSpy: jasmine.SpyObj<Logger>;
+
+  
+  beforeEach(
+    waitForAsync(() => {
+      bookingServiceSpy = jasmine.createSpyObj('BookingService', ['']);
+      videoHearingsServiceSpy = jasmine.createSpyObj('VideoHearingsService', ['']);
+      routerSpy = jasmine.createSpyObj('Router', ['']);
+      loggerSpy = jasmine.createSpyObj('Logger', ['']);
+
+      TestBed.configureTestingModule({
+        declarations: [AddStaffMemberComponent],
+        providers: [
+          provide(BookingService, bookingServiceSpy),
+          provide(Logger, loggerSpy),
+          provide(Router, routerSpy),
+          provide(VideoHearingsService, videoHearingsServiceSpy),
+        ],
+      })
+        .compileComponents();
     })
-    .compileComponents();
-  });
+  );
 
   beforeEach(() => {
     fixture = TestBed.createComponent(AddStaffMemberComponent);
@@ -19,7 +46,13 @@ describe('AddStaffMemberComponent', () => {
     fixture.detectChanges();
   });
 
-  it('should create', () => {
-    expect(component).toBeTruthy();
+  describe('ngOnInit', () => {
+    it('should initialise the form', () => {
+      component.initialiseForm = spyOn(component, 'initialiseForm');
+
+      component.ngOnInit();
+
+      expect(component.initialiseForm).toHaveBeenCalledTimes(1);
+    });
   });
 });
