+++ AdminWebsite/AdminWebsite/ClientApp/src/app/booking/add-staff-member/add-staff-member.component.ts
@@ -1,70 +1,33 @@
 import { Component, OnInit } from '@angular/core';
-import { FormBuilder, FormControl, FormGroup, Validators } from '@angular/forms';
-import { Constants } from 'src/app/common/constants';
 import { ParticipantModel } from 'src/app/common/model/participant.model';
+import { AddParticipantBaseDirective } from 'src/app/booking/add-participant-base/add-participant-base.component'
+import { Router } from '@angular/router';
+import { BookingService } from 'src/app/services/booking.service';
+import { Logger } from 'src/app/services/logger';
+import { VideoHearingsService } from 'src/app/services/video-hearings.service';
 
 @Component({
   selector: 'app-add-staff-member',
   templateUrl: './add-staff-member.component.html',
   styleUrls: ['./add-staff-member.component.css']
 })
-export class AddStaffMemberComponent implements OnInit {
-  constants = Constants;
-
-  form: FormGroup;
-
-  firstNameControl: FormControl;
-  lastNameControl: FormControl;
-  displayNameControl: FormControl;
-  emailControl: FormControl;
-  phoneControl: FormControl;
-
+export class AddStaffMemberComponent extends AddParticipantBaseDirective implements OnInit {
   private staffMember: ParticipantModel;
 
   constructor(
-    private formBuilder: FormBuilder) { }
+    protected bookingService: BookingService,
+    protected router: Router,
+    protected videoHearingService: VideoHearingsService,
+    protected logger: Logger) {
+      super(bookingService, router, videoHearingService, logger);
+    }
 
   ngOnInit(): void {
     this.staffMember = new ParticipantModel();
-    this.initialiseFormFields();
-  }
-
-  get isEmailInvalid(): boolean {
-    return this.emailControl.invalid && (this.emailControl.dirty || this.emailControl.touched);
+    this.initialiseForm();
   }
   
   saveStaffMember(): void {
     console.log(this.staffMember);
-  }
-
-  private initialiseFormFields() {
-    this.firstNameControl = new FormControl(this.staffMember.first_name, {
-        validators: [Validators.required, Validators.pattern(Constants.TextInputPattern), Validators.maxLength(255)],
-        updateOn: 'blur'
-    });
-    this.lastNameControl = new FormControl(this.staffMember.last_name, {
-        validators: [Validators.required, Validators.pattern(Constants.TextInputPattern), Validators.maxLength(255)],
-        updateOn: 'blur'
-    });
-    this.displayNameControl = new FormControl(this.staffMember.display_name, {
-        validators: [Validators.required, Validators.pattern(Constants.TextInputPattern), Validators.maxLength(255)],
-        updateOn: 'blur'
-    });
-    this.emailControl = new FormControl(this.staffMember.email, {
-        validators: [Validators.required, Validators.pattern(Constants.EmailPattern), Validators.maxLength(255)],
-        updateOn: 'blur'
-    });
-    this.phoneControl = new FormControl(this.staffMember.phone, {
-        validators: [Validators.pattern(Constants.PhonePattern)],
-        updateOn: 'blur'
-    });
-    
-    this.form = this.formBuilder.group({
-      firstNameControl: this.firstNameControl,
-      lastNameControl: this.lastNameControl,
-      displayNameControl: this.displayNameControl,
-      emailControl: this.emailControl,
-      phoneControl: this.phoneControl,
-    });
   }
 }
