+++ AdminWebsite/AdminWebsite/ClientApp/src/app/testing/helpers/jasmine.helpers.ts
@@ -1,0 +1,3 @@
+export const provide = (type, value) => {
+  return { provide: type, useValue: value };
+};
